Perbarui Fitur Pelanggaran SDM, Tambah & Ubah Sanksi

+++ app/Http/Controllers/SDM/Pelanggaran.php
@@ -5,12 +5,8 @@ namespace App\Http\Controllers\SDM;
 use Illuminate\Support\Arr;
 use App\Tambahan\FungsiStatis;
 use Illuminate\Validation\Rule;
-use App\Tambahan\CustomValueBinder;
 use App\Http\Controllers\SDM\Berkas;
-use PhpOffice\PhpSpreadsheet\Cell\Cell;
-use PhpOffice\PhpSpreadsheet\IOFactory;
-use PhpOffice\PhpWord\TemplateProcessor;
-use PhpOffice\PhpSpreadsheet\Spreadsheet;
+
 
 class Pelanggaran
 {
@@ -67,17 +63,11 @@ class Pelanggaran
 
         $database = $app->db;
 
-        $kontrak = $database->query()->select('penempatan_uuid', 'penempatan_no_absen', 'penempatan_posisi', 'penempatan_lokasi', 'penempatan_kontrak', 'penempatan_mulai', 'penempatan_selesai', 'penempatan_ke', 'penempatan_keterangan')
-            ->from('penempatans as p1')->where('penempatan_mulai', '=', function ($query) use ($database) {
-                $query->select($database->raw('MAX(penempatan_mulai)'))->from('penempatans as p2')->whereColumn('p1.penempatan_no_absen', 'p2.penempatan_no_absen');
-            });
+        $kontrak = $this->dataKontrak($database);
 
-        $sanksi = $database->query()->select('sanksi_no_absen', 'sanksi_jenis', 'sanksi_lap_no', 'sanksi_selesai')
-            ->from('sanksisdms as p1')->where('sanksi_selesai', '=', function ($query) use ($database) {
-                $query->select($database->raw('MAX(sanksi_selesai)'))->from('sanksisdms as p2')->whereColumn('p1.sanksi_no_absen', 'p2.sanksi_no_absen');
-            });
+        $sanksi = $this->dataSanksi($database);
 
-        $cari = $this->dataDasar()->addSelect('a.sdm_uuid as langgar_tsdm_uuid', 'a.sdm_nama as langgar_tsdm_nama', 'a.sdm_tgl_berhenti as langgar_tsdm_tgl_berhenti', 'kontrak_t.penempatan_lokasi as langgar_tlokasi', 'kontrak_t.penempatan_posisi as langgar_tposisi', 'kontrak_t.penempatan_kontrak as langgar_tkontrak', 'b.sdm_uuid as langgar_psdm_uuid', 'b.sdm_nama as langgar_psdm_nama', 'b.sdm_tgl_berhenti as langgar_psdm_tgl_berhenti', 'kontrak_p.penempatan_lokasi as langgar_plokasi', 'kontrak_p.penempatan_posisi as langgar_pposisi', 'kontrak_p.penempatan_kontrak as langgar_pkontrak', 'sanksilama.sanksi_jenis as sanksi_aktif_sebelumnya', 'sanksilama.sanksi_lap_no as lap_no_sebelumnya', 'sanksilama.sanksi_selesai as sanksi_selesai_sebelumnya', 'sanksisdms.sanksi_uuid as final_sanksi_uuid', 'sanksisdms.sanksi_jenis as final_sanksi_jenis', 'sanksisdms.sanksi_mulai as final_sanksi_mulai', 'sanksisdms.sanksi_selesai as final_sanksi_selesai', 'sanksisdms.sanksi_tambahan as final_sanksi_tambahan', 'sanksisdms.sanksi_keterangan as final_sanksi_keterangan')
+        $cari = $this->dataDasar($database)->addSelect('a.sdm_uuid as langgar_tsdm_uuid', 'a.sdm_nama as langgar_tsdm_nama', 'a.sdm_tgl_berhenti as langgar_tsdm_tgl_berhenti', 'kontrak_t.penempatan_lokasi as langgar_tlokasi', 'kontrak_t.penempatan_posisi as langgar_tposisi', 'kontrak_t.penempatan_kontrak as langgar_tkontrak', 'b.sdm_uuid as langgar_psdm_uuid', 'b.sdm_nama as langgar_psdm_nama', 'b.sdm_tgl_berhenti as langgar_psdm_tgl_berhenti', 'kontrak_p.penempatan_lokasi as langgar_plokasi', 'kontrak_p.penempatan_posisi as langgar_pposisi', 'kontrak_p.penempatan_kontrak as langgar_pkontrak', 'sanksilama.sanksi_jenis as sanksi_aktif_sebelumnya', 'sanksilama.sanksi_lap_no as lap_no_sebelumnya', 'sanksilama.sanksi_selesai as sanksi_selesai_sebelumnya', 'sanksisdms.sanksi_uuid as final_sanksi_uuid', 'sanksisdms.sanksi_jenis as final_sanksi_jenis', 'sanksisdms.sanksi_mulai as final_sanksi_mulai', 'sanksisdms.sanksi_selesai as final_sanksi_selesai', 'sanksisdms.sanksi_tambahan as final_sanksi_tambahan', 'sanksisdms.sanksi_keterangan as final_sanksi_keterangan')
             ->join('sdms as a', 'langgar_no_absen', '=', 'a.sdm_no_absen')
             ->join('sdms as b', 'langgar_pelapor', '=', 'b.sdm_no_absen')
             ->leftJoinSub($kontrak, 'kontrak_t', function ($join) {
@@ -187,9 +177,25 @@ class Pelanggaran
         ];
     }
 
-    public function dataDasar()
+    public function dataDasar($database)
     {
-        return app('db')->query()->select('langgar_uuid', 'langgar_lap_no', 'langgar_no_absen', 'langgar_pelapor', 'langgar_tanggal', 'langgar_status', 'langgar_isi', 'langgar_keterangan')->from('pelanggaransdms');
+        return $database->query()->select('langgar_uuid', 'langgar_lap_no', 'langgar_no_absen', 'langgar_pelapor', 'langgar_tanggal', 'langgar_status', 'langgar_isi', 'langgar_keterangan')->from('pelanggaransdms');
+    }
+
+    public function dataKontrak($database)
+    {
+        return $database->query()->select('penempatan_uuid', 'penempatan_no_absen', 'penempatan_posisi', 'penempatan_lokasi', 'penempatan_kontrak', 'penempatan_mulai', 'penempatan_selesai', 'penempatan_ke', 'penempatan_keterangan')
+            ->from('penempatans as p1')->where('penempatan_mulai', '=', function ($query) use ($database) {
+                $query->select($database->raw('MAX(penempatan_mulai)'))->from('penempatans as p2')->whereColumn('p1.penempatan_no_absen', 'p2.penempatan_no_absen');
+            });
+    }
+
+    public function dataSanksi($database)
+    {
+        return $database->query()->select('sanksi_no_absen', 'sanksi_jenis', 'sanksi_lap_no', 'sanksi_selesai', 'sanksi_mulai')
+            ->from('sanksisdms as p1')->where('sanksi_selesai', '=', function ($query) use ($database) {
+                $query->select($database->raw('MAX(sanksi_selesai)'))->from('sanksisdms as p2')->whereColumn('p1.sanksi_no_absen', 'p2.sanksi_no_absen');
+            });
     }
 
     public function dasarValidasi()
@@ -208,61 +214,6 @@ class Pelanggaran
         ];
     }
 
-    public function formulir($uuid = null)
-    {
-        // $app = app();
-        // $reqs = $app->request;
-        // $pengguna = $reqs->user();
-
-        // abort_unless($pengguna && $uuid && str()->contains($pengguna?->sdm_hak_akses, ['SDM-PENGURUS', 'SDM-MANAJEMEN']), 403, 'Akses dibatasi hanya untuk Pemangku SDM.');
-
-        // set_time_limit(0);
-        // ob_implicit_flush();
-        // ob_end_flush();
-        // header('X-Accel-Buffering: no');
-
-        // echo '<p>Memeriksa formulir.</p>';
-
-        // $lingkupIjin = array_filter(explode(',', $pengguna->sdm_ijin_akses));
-
-        // $permin = $this->dataDasar()->clone()->addSelect('tambahsdm_uuid', 'b.sdm_nama')
-        // ->join('sdms as b', 'tambahsdm_sdm_id', '=', 'b.sdm_no_absen')->where('tambahsdm_uuid', $uuid)->when($lingkupIjin, function ($query,$lingkupIjin) {
-        //     $query->whereIn('tambahsdm_penempatan', $lingkupIjin);
-        // })->first();
-
-        // abort_unless($permin, 404, 'Data Permintaan Tambah SDM tidak ditemukan.');
-
-        // $storage = $app->filesystem;
-
-        // abort_unless($storage->exists("contoh/permintaan-tambah-sdm.docx"), 404, 'Berkas Contoh Formulir Tidak Ditemukan.');
-
-        // $filename = 'permintaan-tambah-sdm-'.$permin->tambahsdm_no.'.docx';
-
-        // $templateProcessor = new TemplateProcessor($app->storagePath('app/contoh/permintaan-tambah-sdm.docx'));
-
-        // echo '<p>Menyiapkan formulir.</p>';
-
-        // $date = $app->date;
-        // $str = str();
-
-        // $templateProcessor->setValues([
-        //     'tambahsdm_no' => $permin->tambahsdm_no,
-        //     'sdm_nama' => $str->limit($permin->sdm_nama, 30),
-        //     'tambahsdm_sdm_id' => $permin->tambahsdm_sdm_id,
-        //     'tambahsdm_posisi' => $str->limit($permin->tambahsdm_posisi, 30),
-        //     'tambahsdm_jumlah' => $permin->tambahsdm_jumlah,
-        //     'tambahsdm_alasan' => $str->limit($permin->tambahsdm_alasan, 100),
-        //     'tambahsdm_tgl_diusulkan' => strtoupper($date->make($permin->tambahsdm_tgl_diusulkan)?->translatedFormat('d F Y')),
-        //     'tambahsdm_tgl_dibutuhkan' => strtoupper($date->make($permin->tambahsdm_tgl_dibutuhkan)?->translatedFormat('d F Y'))
-        // ]);
-
-        // $templateProcessor->saveAs($app->storagePath("app/unduh/{$filename}"));
-
-        // echo '<p>Selesai menyiapkan berkas formulir. <a href="' . $storage->disk('local')->temporaryUrl("unduh/{$filename}", $date->now()->addMinutes(5)) . '">Unduh</a>.</p>';
-
-        // exit();
-    }
-
     public function lihat($uuid = null)
     {
         $app = app();
@@ -275,18 +226,11 @@ class Pelanggaran
 
         $database = $app->db;
 
-        $kontrak = $database->query()->select('penempatan_uuid', 'penempatan_no_absen', 'penempatan_posisi', 'penempatan_lokasi', 'penempatan_kontrak', 'penempatan_mulai', 'penempatan_selesai', 'penempatan_ke', 'penempatan_keterangan')
-            ->from('penempatans as p1')->where('penempatan_mulai', '=', function ($query) use ($database) {
-                $query->select($database->raw('MAX(penempatan_mulai)'))->from('penempatans as p2')->whereColumn('p1.penempatan_no_absen', 'p2.penempatan_no_absen');
-            });
+        $kontrak = $this->dataKontrak($database);
 
+        $sanksi = $this->dataSanksi($database);
 
-        $sanksi = $database->query()->select('sanksi_no_absen', 'sanksi_jenis', 'sanksi_lap_no', 'sanksi_selesai')
-            ->from('sanksisdms as p1')->where('sanksi_selesai', '=', function ($query) use ($database) {
-                $query->select($database->raw('MAX(sanksi_selesai)'))->from('sanksisdms as p2')->whereColumn('p1.sanksi_no_absen', 'p2.sanksi_no_absen');
-            });
-
-        $langgar = $this->dataDasar()->addSelect('a.sdm_uuid as langgar_tsdm_uuid', 'a.sdm_nama as langgar_tsdm_nama', 'a.sdm_tgl_berhenti as langgar_tsdm_tgl_berhenti', 'kontrak_t.penempatan_lokasi as langgar_tlokasi', 'kontrak_t.penempatan_posisi as langgar_tposisi', 'kontrak_t.penempatan_kontrak as langgar_tkontrak', 'b.sdm_uuid as langgar_psdm_uuid', 'b.sdm_nama as langgar_psdm_nama', 'b.sdm_tgl_berhenti as langgar_psdm_tgl_berhenti', 'kontrak_p.penempatan_lokasi as langgar_plokasi', 'kontrak_p.penempatan_posisi as langgar_pposisi', 'kontrak_p.penempatan_kontrak as langgar_pkontrak', 'sanksilama.sanksi_jenis as sanksi_aktif_sebelumnya', 'sanksilama.sanksi_lap_no as lap_no_sebelumnya', 'sanksilama.sanksi_selesai as sanksi_selesai_sebelumnya', 'sanksisdms.sanksi_uuid as final_sanksi_uuid', 'sanksisdms.sanksi_jenis as final_sanksi_jenis', 'sanksisdms.sanksi_mulai as final_sanksi_mulai', 'sanksisdms.sanksi_selesai as final_sanksi_selesai', 'sanksisdms.sanksi_tambahan as final_sanksi_tambahan', 'sanksisdms.sanksi_keterangan as final_sanksi_keterangan')
+        $langgar = $this->dataDasar($database)->addSelect('a.sdm_uuid as langgar_tsdm_uuid', 'a.sdm_nama as langgar_tsdm_nama', 'a.sdm_tgl_berhenti as langgar_tsdm_tgl_berhenti', 'kontrak_t.penempatan_lokasi as langgar_tlokasi', 'kontrak_t.penempatan_posisi as langgar_tposisi', 'kontrak_t.penempatan_kontrak as langgar_tkontrak', 'b.sdm_uuid as langgar_psdm_uuid', 'b.sdm_nama as langgar_psdm_nama', 'b.sdm_tgl_berhenti as langgar_psdm_tgl_berhenti', 'kontrak_p.penempatan_lokasi as langgar_plokasi', 'kontrak_p.penempatan_posisi as langgar_pposisi', 'kontrak_p.penempatan_kontrak as langgar_pkontrak', 'sanksilama.sanksi_jenis as sanksi_aktif_sebelumnya', 'sanksilama.sanksi_lap_no as lap_no_sebelumnya', 'sanksilama.sanksi_mulai as sanksi_mulai_sebelumnya', 'sanksilama.sanksi_selesai as sanksi_selesai_sebelumnya', 'sanksisdms.sanksi_uuid as final_sanksi_uuid', 'sanksisdms.sanksi_jenis as final_sanksi_jenis', 'sanksisdms.sanksi_mulai as final_sanksi_mulai', 'sanksisdms.sanksi_selesai as final_sanksi_selesai', 'sanksisdms.sanksi_tambahan as final_sanksi_tambahan', 'sanksisdms.sanksi_keterangan as final_sanksi_keterangan')
             ->join('sdms as a', 'langgar_no_absen', '=', 'a.sdm_no_absen')
             ->join('sdms as b', 'langgar_pelapor', '=', 'b.sdm_no_absen')
             ->leftJoinSub($kontrak, 'kontrak_t', function ($join) {
@@ -332,7 +276,7 @@ class Pelanggaran
 
             $database = $app->db;
 
-            $hitungLaporan = $this->dataDasar()->addSelect($database->raw('COUNT(DISTINCT langgar_lap_no) jml_lap'))->whereYear('langgar_dibuat', date('Y'))->whereMonth('langgar_dibuat', date('m'))->groupBy('langgar_lap_no');
+            $hitungLaporan = $this->dataDasar($database)->addSelect($database->raw('COUNT(DISTINCT langgar_lap_no) jml_lap'))->whereYear('langgar_dibuat', date('Y'))->whereMonth('langgar_dibuat', date('m'))->groupBy('langgar_lap_no');
 
             $hitungNomor = $database->query()->select('nolap.jml_lap')->fromSub($hitungLaporan, 'nolap')->sum('jml_lap');
 
@@ -437,17 +381,24 @@ class Pelanggaran
 
         $database = $app->db;
 
-        $kontrak = $database->query()->select('penempatan_uuid', 'penempatan_no_absen', 'penempatan_posisi', 'penempatan_lokasi', 'penempatan_kontrak', 'penempatan_mulai', 'penempatan_selesai', 'penempatan_ke', 'penempatan_keterangan')
-            ->from('penempatans as p1')->where('penempatan_mulai', '=', function ($query) use ($database) {
-                $query->select($database->raw('MAX(penempatan_mulai)'))->from('penempatans as p2')->whereColumn('p1.penempatan_no_absen', 'p2.penempatan_no_absen');
-            });
+        $kontrak = $this->dataKontrak($database);
 
-        $langgar = $this->dataDasar()
+        $sanksi = $this->dataSanksi($database);
+
+        $langgar = $this->dataDasar($database)->addSelect('sanksisdms.sanksi_uuid as final_sanksi_uuid')
+            ->join('sdms as a', 'langgar_no_absen', '=', 'a.sdm_no_absen')
+            ->join('sdms as b', 'langgar_pelapor', '=', 'b.sdm_no_absen')
             ->leftJoinSub($kontrak, 'kontrak_t', function ($join) {
                 $join->on('langgar_no_absen', '=', 'kontrak_t.penempatan_no_absen');
             })
             ->leftJoinSub($kontrak, 'kontrak_p', function ($join) {
                 $join->on('langgar_pelapor', '=', 'kontrak_p.penempatan_no_absen');
+            })
+            ->leftJoinSub($sanksi, 'sanksilama', function ($join) {
+                $join->on('langgar_no_absen', '=', 'sanksilama.sanksi_no_absen')->on('sanksilama.sanksi_selesai', '>=', 'langgar_tanggal')->on('langgar_lap_no', '!=', 'sanksilama.sanksi_lap_no');
+            })
+            ->leftJoin('sanksisdms', function ($join) {
+                $join->on('langgar_no_absen', '=', 'sanksisdms.sanksi_no_absen')->on('langgar_lap_no', '=', 'sanksisdms.sanksi_lap_no');
             })
             ->when($lingkupIjin, function ($query) use ($lingkupIjin) {
                 $query->where(function ($group) use ($lingkupIjin) {
@@ -497,13 +448,21 @@ class Pelanggaran
 
             $validasi->validate();
 
+            $session = $reqs->session();
+            $perujuk = $session->get('tautan_perujuk');
+            $redirect = $app->redirect;
+            $kesalahan = 'Laporan yang sudah diberikan sanksi tidak dapat dibatalkan.';
+
+            if ($validasi->safe()->only('langgar_status')['langgar_status'] == 'DIBATALKAN' && $langgar->final_sanksi_uuid) {
+                return $perujuk ? $redirect->to($perujuk)->withErrors($kesalahan) : $redirect->route('sdm.pelanggaran.data')->withErrors($kesalahan);
+            }
+
             $data = $validasi->safe()->except('berkas_laporan');
 
-            $app->db->table('pelanggaransdms')->where('langgar_uuid', $uuid)->update($data);
+            $database->table('pelanggaransdms')->where('langgar_uuid', $uuid)->update($data);
 
             $fungsiStatis->hapusCacheSDMUmum();
             $pesan = $fungsiStatis->statusBerhasil();
-            $session = $reqs->session();
 
             if ($reqs->header('X-Minta-Javascript', false)) {
                 $session->now('pesan', $pesan);
@@ -517,9 +476,6 @@ class Pelanggaran
             if ($berkas) {
                 $berkas->storeAs('sdm/pelanggaran/berkas', $nomorLaporan . '.pdf');
             }
-
-            $perujuk = $session->get('tautan_perujuk');
-            $redirect = $app->redirect;
 
             return $perujuk ? $redirect->to($perujuk)->with('pesan', $pesan) : $redirect->route('sdm.pelanggaran.data')->with('pesan', $pesan);
         }
